+++ src/library_pthread.js
@@ -432,6 +432,18 @@ var LibraryPThread = {
           worker.onerror = function(e) {
             err('pthread sent an error! ' + e.filename + ':' + e.lineno + ': ' + e.message);
           };
+
+          if (ENVIRONMENT_HAS_NODE) {
+            worker.on('message', function(data) {
+              worker.onmessage({ data: data });
+            });
+            worker.on('error', function(data) {
+              worker.onerror(data.err);
+            });
+            worker.on('exit', function(data) {
+              console.log('worker exited - TODO: update the worker queue?');
+            });
+          }
         }(worker));
       }  // for each worker
     },
@@ -503,6 +515,7 @@ var LibraryPThread = {
     if (ENVIRONMENT_IS_PTHREAD) throw 'Internal Error! _spawn_thread() can only ever be called from main application thread!';
 
     var worker = PThread.getNewWorker();
+
     if (worker.pthread !== undefined) throw 'Internal error!';
     if (!threadParams.pthread_ptr) throw 'Internal error, no pthread ptr!';
     PThread.runningWorkers.push(worker);
@@ -944,6 +957,11 @@ var LibraryPThread = {
     else PThread.threadExit(status);
 #if WASM_BACKEND
     // pthread_exit is marked noReturn, so we must not return from it.
+    if (ENVIRONMENT_HAS_NODE) {
+      // exit the pthread properly on node, as a normal JS exception will halt
+      // the entire application.
+      process.exit(status);
+    }
     throw 'pthread_exit';
 #endif
   },