+++ ChangeLog.md
@@ -25,6 +25,7 @@ Current Trunk
  - Add `pthread_tryjoin_np`, which is a POSIX API similar to `pthread_join`
    but without blocking.
  - New function emscripten_has_asyncify()
+ - Add support for pthreads in Node.js, using Node Workers. See #9745
 
 v1.39.1: 10/30/2019
 -------------------